+++ demos/kotlin/tutorial/app/src/main/java/com/example/open_gopro_tutorial/tutorials/Tutorial3ParseBleTlvResponses.kt
@@ -7,7 +7,12 @@ import androidx.annotation.RequiresPermission
 import com.example.open_gopro_tutorial.AppContainer
 import com.example.open_gopro_tutorial.DataStore
 import com.example.open_gopro_tutorial.network.BleEventListener
-import com.example.open_gopro_tutorial.util.*
+import com.example.open_gopro_tutorial.util.GoProUUID
+import com.example.open_gopro_tutorial.util.decodeToString
+import com.example.open_gopro_tutorial.util.prettyJson
+import com.example.open_gopro_tutorial.util.toHexString
+import com.example.open_gopro_tutorial.util.toInt
+import com.example.open_gopro_tutorial.util.toJsonElement
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.Channel
@@ -15,11 +20,11 @@ import kotlinx.coroutines.launch
 import kotlinx.serialization.encodeToString
 import timber.log.Timber
 import java.io.File
-import java.util.*
+import java.util.UUID
 import kotlin.properties.Delegates.notNull
 
 @OptIn(ExperimentalUnsignedTypes::class)
-sealed class Response<T > {
+sealed class Response(val uuid: GoProUUID) {
     private enum class Header(val value: UByte) {
         GENERAL(0b00U), EXT_13(0b01U), EXT_16(0b10U), RESERVED(0b11U);
 
@@ -40,18 +45,10 @@ sealed class Response<T > {
     }
 
     private var bytesRemaining = 0
-    protected var packet = ubyteArrayOf()
-    abstract val data: T
-    var id by notNull<Int>()
-        protected set
-    var status by notNull<Int>()
+    var rawBytes = ubyteArrayOf()
         protected set
 
     val isReceived get() = bytesRemaining == 0
-    var isParsed = false
-        protected set
-
-    override fun toString() = prettyJson.encodeToString(data.toJsonElement())
 
     fun accumulate(data: UByteArray) {
         var buf = data
@@ -60,72 +57,65 @@ sealed class Response<T > {
             buf = buf.drop(1).toUByteArray() // Pop the header byte
         } else {
             // This is a new packet so start with empty array
-            packet = ubyteArrayOf()
+            rawBytes = ubyteArrayOf()
             when (Header.fromValue((buf.first() and Mask.Header.value).toInt() shr 5)) {
                 Header.GENERAL -> {
                     bytesRemaining = buf[0].and(Mask.GenLength.value).toInt()
                     buf = buf.drop(1).toUByteArray()
                 }
+
                 Header.EXT_13 -> {
                     bytesRemaining = ((buf[0].and(Mask.Ext13Byte0.value)
                         .toLong() shl 8) or buf[1].toLong()).toInt()
                     buf = buf.drop(2).toUByteArray()
                 }
+
                 Header.EXT_16 -> {
                     bytesRemaining = ((buf[1].toLong() shl 8) or buf[2].toLong()).toInt()
                     buf = buf.drop(3).toUByteArray()
                 }
+
                 Header.RESERVED -> {
                     throw Exception("Unexpected RESERVED header")
                 }
             }
         }
         // Accumulate the payload now that headers are handled and dropped
-        packet += buf
+        rawBytes += buf
         bytesRemaining -= buf.size
-        Timber.i("Received packet of length ${buf.size}. $bytesRemaining bytes remaining")
+        Timber.d("Received packet of length ${buf.size}. $bytesRemaining bytes remaining")
 
         if (bytesRemaining < 0) {
             throw Exception("Unrecoverable parsing error. Received too much data.")
         }
     }
 
-    abstract fun parse()
+    open class Tlv(uuid: GoProUUID) : Response(uuid) {
+        var payload = ubyteArrayOf()
+            private set
+        var id by notNull<Int>()
+            private set
+        var status by notNull<Int>()
+            private set
 
-    class Complex : Response<MutableList<UByteArray>>() {
-        override val data: MutableList<UByteArray> = mutableListOf()
-
-        override fun parse() {
+        open fun parse() {
             require(isReceived)
-            // Parse header bytes
-            id = packet[0].toInt()
-            status = packet[1].toInt()
-            var buf = packet.drop(2)
-            // Parse remaining packet
-            while (buf.isNotEmpty()) {
-                // Get each parameter's ID and length
-                val paramLen = buf[0].toInt()
-                buf = buf.drop(1)
-                // Get the parameter's value
-                val paramVal = buf.take(paramLen)
-                // Store in data list
-                data += paramVal.toUByteArray()
-                // Advance the buffer for continued parsing
-                buf = buf.drop(paramLen)
-            }
-            isParsed = true
+            id = rawBytes[0].toInt()
+            status = rawBytes[1].toInt()
+            // Store remainder as payload for further parsing later
+            payload = rawBytes.drop(2).toUByteArray()
         }
+
+        override fun toString(): String = "ID: $id, Status: $status, Payload: ${payload.toHexString()}"
     }
 
-    class Query : Response<MutableMap<UByte, UByteArray>>() {
-        override val data: MutableMap<UByte, UByteArray> = mutableMapOf()
+    class Query(uuid: GoProUUID) : Tlv(uuid) {
+        val data: MutableMap<UByte, UByteArray> = mutableMapOf()
+        override fun toString() = prettyJson.encodeToString(data.toJsonElement())
 
         override fun parse() {
-            require(isReceived)
-            id = packet[0].toInt()
-            status = packet[1].toInt()
-            // Parse remaining packet
-            var buf = packet.drop(2)
+            super.parse()
+            var buf = payload.toList()
             while (buf.isNotEmpty()) {
                 // Get each parameter's ID and length
                 val paramId = buf[0]
@@ -138,57 +128,124 @@ sealed class Response<T > {
                 // Advance the buffer for continued parsing
                 buf = buf.drop(paramLen)
             }
-            isParsed = true
         }
     }
 
-    companion object {
-        fun fromUuid(uuid: GoProUUID): Response<*> =
-            when (uuid) {
-                GoProUUID.CQ_COMMAND_RSP -> Complex()
-                GoProUUID.CQ_QUERY_RSP -> Query()
-                else -> throw Exception("Not supported")
-            }
-    }
+    class Unknown(uuid: GoProUUID): Response(uuid) {}
 
+    companion object {
+        fun muxByUuid(uuid: GoProUUID) : Response {
+            return when (uuid) {
+                GoProUUID.CQ_SETTING_RSP, GoProUUID.CQ_COMMAND_RSP -> Tlv(uuid)
+                GoProUUID.CQ_QUERY_RSP -> Query(uuid)
+                else -> Unknown(uuid)
+            }
+        }
+    }
 }
 
 @OptIn(ExperimentalUnsignedTypes::class)
-class Tutorial3ParseBleTlvResponses(number: Int, name: String, prerequisites: List<Prerequisite>) :
+data class OpenGoProVersion(val minor: Int, val major: Int) {
+    companion object {
+        fun fromBytes(data: UByteArray): OpenGoProVersion {
+            var buf = data.toUByteArray()
+            val majorLen = buf[0].toInt()
+            buf = buf.drop(1).toUByteArray()
+            val major = buf.take(majorLen).toInt()
+            buf = buf.drop(1).toUByteArray()
+            val minorLen = buf[0].toInt()
+            buf = buf.drop(1).toUByteArray()
+            val minor = buf.take(minorLen).toInt()
+            return OpenGoProVersion(minor, major)
+        }
+    }
+}
+
+@OptIn(ExperimentalUnsignedTypes::class)
+data class HardwareInfo(
+    val modelNumber: Int,
+    val modelName: String,
+    val firmwareVersion: String,
+    val serialNumber: String,
+    val apSsid: String,
+    val apMacAddress: String
+) {
+    companion object {
+        fun fromBytes(data: UByteArray): HardwareInfo {
+            // Parse header bytes
+            var buf = data.toUByteArray()
+            // Get model number
+            val modelNumLength = buf.first().toInt()
+            buf = buf.drop(1).toUByteArray()
+            val model = buf.take(modelNumLength).toInt()
+            buf = buf.drop(modelNumLength).toUByteArray()
+            // Get model name
+            val modelNameLength = buf.first().toInt()
+            buf = buf.drop(1).toUByteArray()
+            val modelName = buf.take(modelNameLength).decodeToString()
+            buf = buf.drop(modelNameLength).toUByteArray()
+            // Advance past deprecated bytes
+            val deprecatedLength = buf.first().toInt()
+            buf = buf.drop(1).toUByteArray()
+            buf = buf.drop(deprecatedLength).toUByteArray()
+            // Get firmware version
+            val firmwareLength = buf.first().toInt()
+            buf = buf.drop(1).toUByteArray()
+            val firmware = buf.take(firmwareLength).decodeToString()
+            buf = buf.drop(firmwareLength).toUByteArray()
+            // Get serial number
+            val serialLength = buf.first().toInt()
+            buf = buf.drop(1).toUByteArray()
+            val serial = buf.take(serialLength).decodeToString()
+            buf = buf.drop(serialLength).toUByteArray()
+            // Get AP SSID
+            val ssidLength = buf.first().toInt()
+            buf = buf.drop(1).toUByteArray()
+            val ssid = buf.take(ssidLength).decodeToString()
+            buf = buf.drop(ssidLength).toUByteArray()
+            // Get MAC Address
+            val macLength = buf.first().toInt()
+            buf = buf.drop(1).toUByteArray()
+            val mac = buf.take(macLength).decodeToString()
+
+            return HardwareInfo(model, modelName, firmware, serial, ssid, mac)
+        }
+    }
+}
+
+@OptIn(ExperimentalUnsignedTypes::class)
+class Tutorial3ParseBleTlvResponses(
+    number: Int,
+    name: String,
+    prerequisites: List<Prerequisite>
+) :
     Tutorial(number, name, prerequisites) {
-    private val receivedResponse: Channel<Response<*>> = Channel()
-    private var response: Response<*>? = null
+    private val receivedResponses: Channel<Response> = Channel()
+    private val responsesByUuid = GoProUUID.mapByUuid { Response.muxByUuid(it) }
 
     @OptIn(ExperimentalUnsignedTypes::class)
-    private fun tlvResponseNotificationHandler(characteristic: UUID, data: UByteArray) {
-        GoProUUID.fromUuid(characteristic)?.let { uuid ->
-            // If response is currently empty, create a new one
-            response = response ?: Response.fromUuid(uuid)
-        }
-            ?: return // We don't care about non-GoPro characteristics (i.e. the BT Core Battery service)
+    private fun notificationHandler(characteristic: UUID, data: UByteArray) {
+        // Get the UUID (assuming it is a GoPro UUID)
+        val uuid = GoProUUID.fromUuid(characteristic) ?: return
+        Timber.d("Received response on $uuid")
 
-        Timber.d("Received response on $characteristic: ${data.toHexString()}")
-
-        response?.let { rsp ->
-            rsp.accumulate(data)
-            if (rsp.isReceived) {
-                rsp.parse()
-
-                // If the response has success status...
-                if (rsp.status == 0) Timber.i("Received the expected successful response")
-                else Timber.i("Received unexpected response")
-
-                // Notify the command sender the the procedure is complete
-                response = null // Clear for next command
-                CoroutineScope(Dispatchers.IO).launch { receivedResponse.send(rsp) }
+        responsesByUuid[uuid]?.let { response ->
+            response.accumulate(data)
+            if (response.isReceived) {
+                if (uuid == GoProUUID.CQ_COMMAND_RSP) {
+                    CoroutineScope(Dispatchers.IO).launch { receivedResponses.send(response) }
+                } else {
+                    Timber.e("Unexpected Response")
+                }
+                responsesByUuid[uuid] = Response.muxByUuid(uuid)
             }
-        } ?: throw Exception("This should be impossible")
+        }
     }
 
     @OptIn(ExperimentalUnsignedTypes::class)
     private val bleListeners by lazy {
         BleEventListener().apply {
-            onNotification = ::tlvResponseNotificationHandler
+            onNotification = ::notificationHandler
         }
     }
 
@@ -203,26 +260,27 @@ class Tutorial3ParseBleTlvResponses(number: Int, name: String, prerequisites: Li
         ble.registerListener(goproAddress, bleListeners)
 
         Timber.i("Getting the Open GoPro version")
-        val getVersion = ubyteArrayOf(0x01U, 0x51U)
-        ble.writeCharacteristic(goproAddress, GoProUUID.CQ_COMMAND.uuid, getVersion)
-        val version = receivedResponse.receive() as Response.Complex // Wait to receive response
-        val major = version.data[0].first().toInt()
-        val minor = version.data[1].first().toInt()
-        Timber.i("Got the Open GoPro version successfully: $major.$minor")
+        val versionRequest = ubyteArrayOf(0x01U, 0x51U)
+        ble.writeCharacteristic(goproAddress, GoProUUID.CQ_COMMAND.uuid, versionRequest)
+        // Wait to receive response the parse it
+        var tlvResponse = receivedResponses.receive() as Response.Tlv
+        tlvResponse.parse()
+        Timber.i("Received response: $tlvResponse")
+        val version = OpenGoProVersion.fromBytes(tlvResponse.payload)
+        Timber.i("Got the Open GoPro version successfully: ${version.major}.${version.minor}")
 
-        Timber.i("Getting the camera's settings")
-        val getCameraSettings = ubyteArrayOf(0x01U, 0x12U)
-        ble.writeCharacteristic(goproAddress, GoProUUID.CQ_QUERY.uuid, getCameraSettings)
-        val settings = receivedResponse.receive()
-        Timber.i("Got the camera's settings successfully")
-        Timber.i(settings.toString())
-
-        Timber.i("Getting the camera's statuses")
-        val getCameraStatuses = ubyteArrayOf(0x01U, 0x13U)
-        ble.writeCharacteristic(goproAddress, GoProUUID.CQ_QUERY.uuid, getCameraStatuses)
-        val statuses = receivedResponse.receive()
-        Timber.i("Got the camera's statuses successfully")
-        Timber.i(statuses.toString())
+        Timber.i("Getting the Hardware Info")
+        val hardwareInfoRequest = ubyteArrayOf(0x01U, 0x3CU)
+        ble.writeCharacteristic(
+            goproAddress,
+            GoProUUID.CQ_COMMAND.uuid,
+            hardwareInfoRequest
+        )
+        // Wait to receive response the parse it
+        tlvResponse = receivedResponses.receive() as Response.Tlv
+        tlvResponse.parse()
+        val hardwareInfo = HardwareInfo.fromBytes(tlvResponse.payload)
+        Timber.i("Got the Hardware Info successfully: $hardwareInfo")
 
         // Other tutorials will use their own notification handler so unregister ours now
         ble.unregisterListener(bleListeners)