+++ demos/python/tutorial/tutorial_modules/tutorial_4_ble_queries/ble_query_register_resolution_value_updates.py
@@ -2,95 +2,86 @@
 # This copyright was auto-generated on Wed, Sep  1, 2021  5:06:00 PM
 
 import sys
-import enum
 import asyncio
 import argparse
-from typing import Optional
 
 from bleak import BleakClient
 from bleak.backends.characteristic import BleakGATTCharacteristic
 
-from tutorial_modules import GOPRO_BASE_UUID, connect_ble, Response
+from tutorial_modules import GoProUuid, connect_ble, QueryResponse, Resolution
 
 from tutorial_modules import logger
 
 
-# Note that this may change based on the Open GoPro version!
-class Resolution(enum.Enum):
-    RES_4K = 1
-    RES_2_7K = 4
-    RES_2_7K_4_3 = 6
-    RES_1080 = 9
-    RES_4K_4_3 = 18
-    RES_5K = 24
-
-
-resolution: Resolution
-
-
-async def main(identifier: Optional[str]) -> None:
-    # Synchronization event to wait until notification response is received
-    event = asyncio.Event()
-
-    # UUIDs to write to and receive responses from
-    SETTINGS_REQ_UUID = GOPRO_BASE_UUID.format("0074")
-    SETTINGS_RSP_UUID = GOPRO_BASE_UUID.format("0075")
-    QUERY_REQ_UUID = GOPRO_BASE_UUID.format("0076")
-    QUERY_RSP_UUID = GOPRO_BASE_UUID.format("0077")
-
+async def main(identifier: str | None) -> None:
     RESOLUTION_ID = 2
 
     client: BleakClient
-    response = Response()
+    responses_by_uuid = GoProUuid.dict_by_uuid(QueryResponse)
+    received_responses: asyncio.Queue[QueryResponse] = asyncio.Queue()
 
-    def notification_handler(characteristic: BleakGATTCharacteristic, data: bytes) -> None:
-        logger.info(f'Received response at handle {characteristic.handle}: {data.hex(":")}')
+    query_request_uuid = GoProUuid.QUERY_REQ_UUID
+    query_response_uuid = GoProUuid.QUERY_RSP_UUID
+    setting_request_uuid = GoProUuid.SETTINGS_REQ_UUID
+    setting_response_uuid = GoProUuid.SETTINGS_RSP_UUID
 
+    async def notification_handler(characteristic: BleakGATTCharacteristic, data: bytearray) -> None:
+        uuid = GoProUuid(client.services.characteristics[characteristic.handle].uuid)
+        logger.info(f'Received response at {uuid}: {data.hex(":")}')
+
+        response = responses_by_uuid[uuid]
         response.accumulate(data)
 
         # Notify the writer if we have received the entire response
         if response.is_received:
-            response.parse()
-
             # If this is query response, it must contain a resolution value
-            if client.services.characteristics[characteristic.handle].uuid == QUERY_RSP_UUID:
-                global resolution
-                resolution = Resolution(response.data[RESOLUTION_ID][0])
+            if uuid is query_response_uuid:
+                logger.info("Received the Resolution Query response")
+                await received_responses.put(response)
             # If this is a setting response, it will just show the status
-            elif client.services.characteristics[characteristic.handle].uuid == SETTINGS_RSP_UUID:
-                logger.info("Command sent successfully")
+            elif uuid is setting_response_uuid:
+                logger.info("Received Set Setting command response.")
+                await received_responses.put(response)
             # Anything else is unexpected. This shouldn't happen
             else:
                 logger.error("Unexpected response")
-
-            # Notify writer that the procedure is complete
-            event.set()
+            # Reset the per-uuid response
+            responses_by_uuid[uuid] = QueryResponse(uuid)
 
     client = await connect_ble(notification_handler, identifier)
 
     # Register for updates when resolution value changes
     logger.info("Registering for resolution updates")
-    event.clear()
-    await client.write_gatt_char(QUERY_REQ_UUID, bytearray([0x02, 0x52, RESOLUTION_ID]), response=True)
-    await event.wait()  # Wait to receive the notification response
+    request = bytes([0x02, 0x52, RESOLUTION_ID])
+    logger.debug(f"Writing to {query_request_uuid}: {request.hex(':')}")
+    await client.write_gatt_char(query_request_uuid.value, request, response=True)
+    # Wait to receive the notification response
+    response = await received_responses.get()
+    response.parse()
     logger.info("Successfully registered for resolution value updates.")
+    resolution = Resolution(response.data[RESOLUTION_ID][0])
     logger.info(f"Resolution is currently {resolution}")
 
     # Write to command request BleUUID to change the video resolution (either to 1080 or 2.7K)
     new_resolution = Resolution.RES_2_7K if resolution is Resolution.RES_1080 else Resolution.RES_1080
     logger.info(f"Changing the resolution to {new_resolution}...")
-    event.clear()
-    await client.write_gatt_char(
-        SETTINGS_REQ_UUID, bytearray([0x03, 0x02, 0x01, new_resolution.value]), response=True
-    )
-    await event.wait()  # Wait to receive the notification response
-    logger.info("Successfully changed the resolution")
+    request = bytes([0x03, 0x02, 0x01, new_resolution.value])
+    logger.debug(f"Writing to {setting_request_uuid}: {request.hex(':')}")
+    await client.write_gatt_char(setting_request_uuid.value, request, response=True)
+    # Wait to receive the notification response
+    response = await received_responses.get()
+    response.parse()
+    # Ensure the setting was successful
+    assert response.status == 0x00
 
     # Let's verify we got the update
-    while resolution is not new_resolution:
-        event.clear()
-        await event.wait()
-    logger.info(f"Resolution is now {resolution}")
+    logger.info("Waiting to receive new resolution")
+    while resolution is not new_resolution and (response := await received_responses.get()):
+        response.parse()
+        resolution = Resolution(response.data[RESOLUTION_ID][0])
+        logger.info(f"Resolution is currently {resolution}")
+
+    logger.info("Resolution has changed as expected. Exiting...")
 
     await client.disconnect()
 
@@ -110,7 +101,7 @@ if __name__ == "__main__":
 
     try:
         asyncio.run(main(args.identifier))
-    except Exception as e:
+    except Exception as e:  # pylint: disable=broad-exception-caught
         logger.error(e)
         sys.exit(-1)
     else: