+++ demos/python/tutorial/tutorial_modules/tutorial_5_ble_protobuf/proto/__init__.py
@@ -1,0 +1,31 @@
+# __init__.py/Open GoPro, Version 2.0 (C) Copyright 2021 GoPro, Inc. (http://gopro.com/OpenGoPro).
+# This copyright was auto-generated on Wed Mar 27 22:05:49 UTC 2024
+
+from .cohn_pb2 import (
+    ResponseCOHNCert,
+    NotifyCOHNStatus,
+    RequestClearCOHNCert,
+    RequestCOHNCert,
+    RequestCreateCOHNCert,
+    RequestGetCOHNStatus,
+    RequestSetCOHNSetting,
+    EnumCOHNNetworkState,
+    EnumCOHNStatus,
+)
+from .network_management_pb2 import (
+    NotifProvisioningState,
+    NotifStartScanning,
+    ResponseGetApEntries,
+    ResponseConnectNew,
+    RequestConnect,
+    RequestConnectNew,
+    RequestGetApEntries,
+    RequestStartScan,
+    ResponseConnect,
+    ResponseStartScanning,
+    EnumProvisioning,
+    EnumScanning,
+    EnumScanEntryFlags,
+)
+from .response_generic_pb2 import ResponseGeneric, EnumResultGeneric
+from .turbo_transfer_pb2 import RequestSetTurboActive